+++ src/Kinematics.hpp
@@ -27,7 +27,7 @@ namespace shared {
             
             Eigen::MatrixXd getPoseOfLinkN(const double &joint_angle, 
             		                       Eigen::MatrixXd &current_transform, 
-										   unsigned int &n) const;
+										   size_t &n) const;
             
             Eigen::MatrixXd getEndEffectorPose(const std::vector<double> &joint_angles, bool &eigen);
